+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [Tobii Provider] If IsLive is false, now reports reason in ErrorMessage
 - [SpaceMouse provider] IsLive is always true, as HID is always present
 - [MIDI provider] IsLive is always true, as MIDI is always present
+- [TitanOne Provider] IsLive now reflects connected status of device
+- [TitanOne Provider] Reports 0 devices if IsLive is false
 ### Changed 
 ### Deprecated
 ### Removed